+++ src/main/components/home/heroSection.js
@@ -3,9 +3,9 @@ import {
   Link,
 } from 'react-router-dom';
 import Styled from 'styled-components';
-import Button from './button';
+import Button from '../button';
 import heatMap from './table.png';
-import searchJson from '../../config/search.json';
+import searchJson from '../../../config/search.json';
 
 const Hero = Styled.div`
   margin: 0 auto 130px;
