+++ src/main/components/home/infoSection.js
