+++ src/main/components/search/spinner.js
